Move slider to components and add to vector default view
