+++ src/components/Input/NumericInput/NumericSlider/hooks.ts
