+++ src/util/numeric.ts
